+++ src/Games/GamesList/GamesList.container.jsx
@@ -1,22 +1,22 @@
 import { useEffect, useContext } from "react";
 import proptypes from 'prop-types';
-import { useParams } from "react-router-dom";
 import { AppState } from "../../Config/store/state";
 import GamesList from "./GamesList.component";
-import { ConsoleService, GameService, GenreService } from '../../services';
-import { GAME_LIST_OPTIONS, OPERATION_OUTCOME } from "../../utils/constants";
+import { GAME_LIST_OPTIONS } from "../../utils/constants";
 import Spinner from "../../Common/Spinner/Spinner.component";
 
 
 const GamesListContainer = ({
   editGame,
   deleteGame,
+  getWishlistByConsole,
+  getGamesByConsole,
 }) => {
-  const { game, setGamesList, openSnackbar, setConsolesList, setGenresList, setIsLoading, isLoading } = useContext(AppState);
-  const { consoleId } = useParams()
+  const { game, isLoading } = useContext(AppState);
 
   useEffect(() => {
     const { listOption } = game;
+    console.log({listOption})
 
     switch(listOption) {
       case GAME_LIST_OPTIONS.ALPHABET:
@@ -32,47 +32,6 @@ const GamesListContainer = ({
     }
     
   }, [game.listOption]);
-
-  const getGamesByConsole = async () => {
-    try {
-      setIsLoading(true)
-      const gamesResponse = await GameService.getByParams({idConsole: consoleId});
-      const consolesResponse = await ConsoleService.getAll();
-      const genresResponse = await GenreService.getAll();
-      setGamesList(gamesResponse.data || []);
-      setConsolesList(consolesResponse.data || []);
-      setGenresList(genresResponse.data || []);
-    }
-    catch(e){
-      console.log(e)
-      openSnackbar({message: e.message, type: OPERATION_OUTCOME.FAILED})
-    }
-    finally {
-      setIsLoading(false)
-    }
-  }
-
-
-  const getWishlistByConsole = async () => {
-    try {
-      setIsLoading(true)
-      const gamesResponse = await GameService.getWishlistByConsole(consoleId);
-      const consolesResponse = await ConsoleService.getAll();
-      const genresResponse = await GenreService.getAll();
-      setGamesList(gamesResponse.data || []);
-      setConsolesList(consolesResponse.data || []);
-      setGenresList(genresResponse.data || []);
-    }
-    catch(e){
-      console.log(e)
-      openSnackbar({message: e.message, type: OPERATION_OUTCOME.FAILED})
-    }
-    finally {
-      setIsLoading(false)
-    }
-  }
-
-  
 
   return (isLoading ? <Spinner /> :
     <GamesList
